checkers: move is_pure_expression() to utils.py and start working on in-range-instead-of-compare
